+++ src/commands/__init__.py
@@ -4,3 +4,6 @@ from .faq import FaqCommand  # noqa
 from .deck import DeckCommand  # noqa
 from .cimg import CimgCommand  # noqa
 from .overlap import OverlapCommand  # noqa
+from .term import TermCommand  # noqa
+from .rulebook import RulebookCommand  # noqa
+from .help import HelpCommand  # noqa